Have the start state 'on_exit' be called when exit occurs
When the initial start state is exited from some event
to a new stable state it is nice to have its 'on_exit'
callback be triggerred.

The 'on_enter' is not triggered since the initialization
of a machine to a start state is not triggered by an event
and is more along the line of 'priming' the state machine,
not triggering any events that cause a transition to occur.