Allow providing a state-name translation dict
Certain projects use ``None`` as a state (and it is
possible for other projects to use other hashable
objects as states) so make it possible to provide
a renaming dict that can provide an alternative name
mapping.

This allows those projects to translate their states
into something useful on conversion.

Change-Id: I1debeaf0c1d7a7af58376ada50b014cf5de40746