Ensure machine special method(s) include in generated docs
Without explicitly mentioning that it should have its docs
generated sphinx won't generate docs for it, so ensure that
we list __iter__ and __contains__ so that it does get
generated docs.

Also fixes 'default_start_state' docs which were not showing
up since they were on the setter method vs the getter method.

Change-Id: I83d02a3604678d16b16f6da5384ff10543af850f