doc: Avoid duplicate entry warning
Both the 'FiniteMachine' and 'HierarchicalFiniteMachine' in
'automaton.machines' define an 'Effect' attribute. It seems recent
versions of autodoc can't handle this since we're seeing the following
warning:

  automaton/machines.py:docstring of
  automaton.machines.Effect:1:duplicate object description of
  automaton.machines.Effect, other instance in reference/index, use
  :noindex: for one of them

Do as it says and configure 'noindex' for the latter. This means we lose
the ability to link to other members of this class but that's a
necessary compromise since there's no way to exclude an individual
member.

Change-Id: Ic9e50f6b56a278032463357552960e40c2ddf6a7
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 45e5452482f480c0ee6cb324e790fd1c71d4eca7)