[stable-only] Cap setuptools <71.0.0
py38 jobs (on ubuntu-focal) started to fail (see the relevant github
issue [1]) due to recent virtualenv release (20.26.4; which bundles
setuptools).

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

Furthermore, lower-constraints jobs are not supported anymore and are
removed from most of the branches, so let's remove here as well to
unblock the gate.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: Ic8725b894a1cd22c0323476284144b8d1d1a4b21